Fixed how the equip table finds item info
Was only checking the bag and wardrobe 1.
Was doing a read based on an if/else of where the item was located.
Now it just calculates based on the size of each container and the index
given in the equip table.
Still need to make this more generic, remove redundant code.