Added stats box filter if jp midnight had occurred
When the fisher is running, we do not want to re-filter the stats box at
the moment Jp Midnight occurs for possibility of overlapping with a
stats box update. So we set a flag that gets checked when the normal
stats box update occurs.  If the flag is set, the normal update is
omitted and we do a re-filter & reload instead.