Split up the inits
Created Init_Iocaine which loads all of the commands at initial load
time.
The individual command sets are loaded on Init_LoggedIn and when player
data is changed (through the change monitor).
Also made it so that the command manager handles the 3 sub-command
managers instead of needing to call them all individually.