Changed how the active sequence is handled.
Previously we were trying to set the active sequence and keep it active
until it was no longer capable of being called.  This isn't going to
work in cases where items may be involved.
For instance, if we initially have a Prism Powder in the bag and we're
PLD/WAR, we will use it for Invisible since we have no other means of
Invisible.  Once we use it, the set sequence is still on that item, even
if there are no more remaining.  IsCapable would still return true.

This has been changed to no longer keep a set sequence.  We now just
find the best sequence on the fly each time it is required.  Same with
IsCapable and CanPerform.