Added IocaineTextbox
Inherits the normal textbox, but adds our default text when not selected
(gray text), selects the text when entering the textbox and the default
text is shown.
Also has the option to set the text to the current target's Name when
hitting enter from the textbox, or if the target is not "" when the
textbox is created.
On setting the target, an event is fired which can cause the owning form
to save, close, or just set the next control to be focused.