Got the initial footprint working.
Adding a textbox works.
If the option is set, upon creation, if the player's target is not null,
the target is entered into the textbox.  Once the form is drawn, the
focus is then set to the OK button.
This OK button focus may need to be optional if we have multiple
controls.
Still lots to add.