Added default target to ActionSequences
If the target passed in is "", the default target is used instead.
The property is default to "<me>", but can be set on a case by case
basis.
Any actions that do not need a target will ignore the value.
Could need to add more flexibility if actions within the same sequence
require different targets.