Fixed form auto-closing when opened with Enter key
Since the UpDown control had to use the KeyUp and KeyDown events to get
rid of the dinging, and the KeyUp was where the Enter key was detected
(and would save the default data), when the user clicked the button on
the form to create a new node that only had an UpDown control, the
KeyDown opened the DataEntry form and the subsequent KeyUp event would
save the data.
Trapped the KeyUp so that it only fires if the KeyDown event had been
previously fired in the same form.