Fixed ninjutsu load.
The ninja job ID was not being set early enough, nor was it being
checked for before creating the command conditions.  This led to some of
the non-player-usable ninjutsu spells that have job level = 0 to have
the tool conditions created, but there is no entry in the maps for those
spells, so an exception was thrown.