Initial functionality in place.
Loads and runs substitutions from file.
Loads categorization filters and runs.
Does not yet do anything with the strings other than remove them from
the description.
This way we can run filters on items until we find text that we cannot
parse. The user is notified and we can add the filters/substitutions,
reload, and reparse without having to recompile.