Added post processing to handle items by the same name.
Many items will still be ambiguous since we do not or cannot
differentiate.  Such as 'Linkshell'.
The important items of note are REM weapons/shields which have a
different ID for each level, as well as for those with Afterglow and an
additional final upgrade (the 10k plutons upgrade).
These all now have seperate keys in the m_ids map.  This is the map used
to look up item information when the user gives the item by name.  Now
the user can differentiate between various versions of items with the
same name.